requirements.txt
aiogram
google-generativeai
httpx
keystone-engine
termcolor